Move the scraper to the package
This way, packages that depend on us can use GMT maps in their galleries
