+++ doc/conf.py
@@ -6,10 +6,9 @@ import shutil
 import datetime
 import sphinx_rtd_theme
 import sphinx_gallery
-from sphinx_gallery.scrapers import figure_rst
 from sphinx_gallery.sorting import FileNameSortKey, ExplicitOrder
-import pygmt
 from pygmt import __version__, __commit__
+from pygmt.sphinx_gallery import PyGMTScraper
 
 
 extensions = [
@@ -26,8 +25,6 @@ extensions = [
     "sphinx_gallery.gen_gallery",
 ]
 
-nbsphinx_allow_errors = True
-
 # Autosummary pages will be generated by sphinx-autogen instead of sphinx-build
 autosummary_generate = False
 
@@ -40,23 +37,6 @@ intersphinx_mapping = {
     "pandas": ("http://pandas.pydata.org/pandas-docs/stable/", None),
     "xarray": ("http://xarray.pydata.org/en/stable/", None),
 }
-
-
-class PyGMTScraper:
-    def __init__(self):
-        self.seen = set()
-
-    def __call__(self, block, block_vars, gallery_conf):
-        image_names = list()
-        image_path_iterator = block_vars["image_path_iterator"]
-        for fig_name in pygmt.get_figures():
-            if fig_name not in self.seen:
-                self.seen.add(fig_name)
-                fname = image_path_iterator.next()
-                pygmt.get_figures()[fig_name].savefig(fname, transparent=True)
-                image_names.append(fname)
-        return figure_rst(image_names, gallery_conf["src_dir"])
-
 
 sphinx_gallery_conf = {
     # path to your examples scripts
