+++ pygmt/__init__.py
@@ -12,7 +12,7 @@ from ._version import get_versions as _get_versions
 
 # Import modules to make the high-level GMT Python API
 from .session_management import begin as _begin, end as _end
-from .figure import Figure, get_figures
+from .figure import Figure
 from .gridding import surface
 from .modules import info, grdinfo, which
 from . import datasets
